Implement Khatri-Rao operator
Remove row-wise flag from parameter/keyword set. That is, only compute the
column-wise Khatri-Rao product.
